絵文字の追加辞書を移動
